:truck: [MCH] [MRRTF-107] Move Digit class from MCHBase to DataFormatsMCH
