+++ DataFormats/Detectors/MUON/MCH/src/Digit.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "MCHBase/Digit.h"
+#include "DataFormatsMCH/Digit.h"
 #include <cmath>
 
 namespace o2::mch
