+++ Detectors/MUON/MCH/Base/CMakeLists.txt
@@ -11,10 +11,9 @@
 o2_add_library(MCHBase
          SOURCES
 	       src/ClusterBlock.cxx
-	       src/Digit.cxx
            src/PreCluster.cxx
            src/TrackBlock.cxx
-	       PUBLIC_LINK_LIBRARIES ROOT::Core FairRoot::Base FairMQ::FairMQ ms_gsl::ms_gsl)
+	       PUBLIC_LINK_LIBRARIES O2::DataFormatsMCH)
 
 o2_target_root_dictionary(MCHBase
-  HEADERS include/MCHBase/Digit.h include/MCHBase/DecoderError.h)
+  HEADERS include/MCHBase/DecoderError.h)
