+++ Detectors/MUON/MCH/Base/src/MCHBaseLinkDef.h
@@ -17,7 +17,4 @@
 #pragma link C++ namespace o2;
 #pragma link C++ namespace o2::mch;
 
-#pragma link C++ class o2::mch::Digit + ;
-#pragma link C++ class std::vector < o2::mch::Digit > +;
-
 #endif
