+++ Detectors/MUON/MCH/Clustering/include/MCHClustering/ClusterFinderOriginal.h
@@ -26,7 +26,7 @@
 
 #include <TH2D.h>
 
-#include "MCHBase/Digit.h"
+#include "DataFormatsMCH/Digit.h"
 #include "MCHBase/ClusterBlock.h"
 #include "MCHMappingInterface/Segmentation.h"
 #include "MCHPreClustering/PreClusterFinder.h"
