+++ Detectors/MUON/MCH/Raw/Encoder/Digit/digit2json.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "MCHBase/Digit.h"
+#include "DataFormatsMCH/Digit.h"
 #include "MCHRawElecMap/Mapper.h"
 #include <boost/program_options.hpp>
 #include <fmt/format.h>
