+++ Detectors/MUON/MCH/Tracking/src/Track.cxx
@@ -17,7 +17,7 @@
 
 #include <iostream>
 
-#include <FairMQLogger.h>
+#include "Framework/Logger.h"
 
 namespace o2
 {
