+++ Detectors/MUON/MCH/Workflow/src/ClusterFinderOriginalSpec.cxx
@@ -33,7 +33,7 @@
 #include "Framework/Logger.h"
 
 #include "DataFormatsMCH/ROFRecord.h"
-#include "MCHBase/Digit.h"
+#include "DataFormatsMCH/Digit.h"
 #include "MCHBase/PreCluster.h"
 #include "MCHBase/ClusterBlock.h"
 #include "MCHClustering/ClusterFinderOriginal.h"
