+++ Detectors/MUON/MCH/Workflow/src/PreClusterSinkSpec.cxx
@@ -32,7 +32,7 @@
 #include "Framework/Logger.h"
 
 #include "DataFormatsMCH/ROFRecord.h"
-#include "MCHBase/Digit.h"
+#include "DataFormatsMCH/Digit.h"
 #include "MCHBase/PreCluster.h"
 #include "MCHMappingInterface/Segmentation.h"
